OXDEV-3103 deprecate AuthenticationService
